// src/general/types.ts
// eslint-disable-next-line import/no-namespace
import * as THREE from 'three'
import { GLTF } from 'three-stdlib'

export type GLTFResult = GLTF & {
    nodes: {
        Cube004_1: THREE.Mesh
        Cube004_2: THREE.Mesh
        Cube004_3: THREE.Mesh
        Cube004_4: THREE.Mesh
        Cube004_5: THREE.Mesh
        Cube002: THREE.Mesh
        Cube003: THREE.Mesh
        Cube004: THREE.Mesh
        Cube005: THREE.Mesh
        Cube006: THREE.Mesh
        Cube007: THREE.Mesh
        Cylinder010: THREE.Mesh
        Cylinder010_1: THREE.Mesh
        Cube012: THREE.Mesh
        Cube012_1: THREE.Mesh
        Cube012_2: THREE.Mesh
        Cube012_3: THREE.Mesh
        Screen: THREE.Mesh
        Tube_cap_top: THREE.Mesh
        Tube_cap_base: THREE.Mesh
        Marble_disc_bottom: THREE.Mesh
        Marble_disc_top: THREE.Mesh
        Cylinder024: THREE.Mesh
        Cylinder024_1: THREE.Mesh
    }
    materials: {
        ['Material.002']: THREE.MeshStandardMaterial
        metal: THREE.MeshStandardMaterial
        sm1: THREE.MeshStandardMaterial
        speakers: THREE.MeshStandardMaterial
        Black: THREE.MeshStandardMaterial
        pMetal: THREE.MeshStandardMaterial
        p: THREE.MeshStandardMaterial
        ['Screen JBC']: THREE.MeshStandardMaterial
        ['Tube lid']: THREE.MeshStandardMaterial
        Marble: THREE.MeshStandardMaterial
        ['Tube base']: THREE.MeshStandardMaterial
        ['Tube glossy.001']: THREE.MeshStandardMaterial
    }
}